pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "openst"
version = "0.0.6"
description = "The computational pipeline for the Open-ST method."
license = "GPL-2.0"
authors = [
    "Daniel León-Periñán <daniel.leonperinan@mdc-berlin.de>",
    "Nikolaos Karaiskos <nikolaos.karaiskos@mdc-berlin.de>",
]
maintainers = [
    "Daniel León-Periñán <daniel.leonperinan@mdc-berlin.de>",
    "Nikolaos Karaiskos <nikolaos.karaiskos@mdc-berlin.de>",
]
readme = "README.md"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Natural Language :: English",
    "Operating System :: MacOS :: MacOS X",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Visualization",
]


[tool.poetry.dependencies]
python = "^3.8"
anndata = ">=0.7.4"
numpy = ">=1.17.0"
matplotlib = ">=3.4"
pandas = ">=1.0"
scipy = ">=1.4"
h5py = ">=3"
tqdm = "*"
scikit-learn = ">=0.24"
natsort = "*"
joblib = "*"
dask = { version = "*", extras = ["array"] }
dask-image = "*"
kornia = ">=0.7.0"
scikit-image = "==0.19.3"
opencv-python = ">=4.8.0.76"
cellpose = ">=2.2.3"
pyqt5 = "==5.15.9"
pyqt5-qt5 = "==5.15.2"
pyqt5-sip = "==12.12.2"
ome-zarr = ">=0.8.2"
pyqtgraph = ">=0.13.3"

[tool.poetry.urls]
"Documentation" = "https://rajewsky-lab.github.io/openst/"
"Source" = "https://github.com/rajewsky-lab/openst"
"Home-page" = "https://rajewsky-lab.github.io/openst"
"Twitter" = "https://twitter.com/open_sts"

[tool.poetry.scripts]
openst = 'openst.__main__:run_openst'
